remove "(by OpenCtx)" in @-mention menu
We want these to feel native.

Also the heading of the @-mention menu when you've selected a provider now just contains the provider name itself, instead of `Search using PROVIDER provider`, which could be ungrammatical or clunky. We should have a way in the OpenCtx `meta` result to specify the right label here, but for now, this suffices.