Document Code: move tests to separate workspace
Previously, the document code tests ran in the monolithic `example-ts`
workspace. This was problematic because when you added a new "Document
Code" test to example-ts, then it would invalidate the HTTP recordings
for unrelated tests in the `index.test.ts` file, because they use
enhanced symf context that would pull in the newly added files.
Now, we run the test with an isolated `document-code` workspace so that
we can add tests without introducing big unrelated diffs in
`index.test.ts`.