Fixed the keyboard shortcuts for "Accept" and "Show diff". (#1671)
Fixes https://github.com/sourcegraph/cody-issues/issues/379.

IIRC IntelliJ automagically converts Windows keyboard shortcuts that use
Ctrl into Mac OS shortcuts that use Cmd. This somehow causes an
additional mapping to appear, e.g. `Cmd-Shift-'` for `Ctrl-Shift-'`;
however, this additional mapping doesn't actually work.

It seems that, when IntelliJ is launched by `./gradlew runIDE`, the
`$default` keymap happens to be called "Windows" even on Mac OS. At
least this is what I observed in a debug session; in the UI, however,
the keymap that's selected by default is "macOS". I haven't investigated
this in depth, but this looks like a bug in IntelliJ.

Whatever the actual cause, setting
[`replace-all="true"`](https://plugins.jetbrains.com/docs/intellij/plugin-configuration-file.html#idea-plugin__actions__action__keyboard-shortcut)
wherever we define more than one shortcut has solved this problem.

After #1409, there are no places in the code where
`DIFF_SESSION_DATA_KEY` is added to the action data context. As a
consequence, the algorithm for mapping the keyboard shortcut to an
action (which IntelliJ does dynamically every time) decided that this
action is disabled, and removed it from the list of possible action
mappings.

## Test plan

Tested manually.