fix: exclude edit & retry lens from unit test command (#1679)
CONTEXT:
https://sourcegraph.slack.com/archives/C05AGQYD528/p1716914610459299
CLOSE https://github.com/sourcegraph/cody-issues/issues/404

This PR aims to disable the `Edit & Retry` lens for the Unit Test
command to match the behavior in VS Code

This commit introduces a new `commandName` property to the
`DocumentCodeSession`, `EditCodeSession`, and `TestCodeSession` classes.
This property is used to determine if a command is a unit test command
in the `FixupSession` class.

It also modifies the `createAcceptGroup` method in the
`LensGroupFactory` class to exclude the "Edit & Retry" action for unit
test commands.

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://sourcegraph.com/docs/dev/background-information/testing_principles

Why does it matter?

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component:
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests"
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs:
  - "previewed locally"
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI"
  - "locally tested"
-->


1. Run the Unit Test command to verify the `Edit & Retry` lens does not
show up.
2. Manually test the "Accept" group creation for both regular and unit
test commands to verify the correct actions are included.


![image](https://github.com/sourcegraph/jetbrains/assets/68532117/54b87841-a947-4635-a97c-cd960663ccbd)

Confirm the lenses are showing up correctly for other edit commands.