Agent: add support for code actions and "Ask Cody to Fix" command
Previously, there was no way for agent clients to use the "Ask Cody to
Fix" command. This PR fixes this problem by adding support for code
actions. Code actions are the "quickfix" menu when you trigger
option-enter over a diagnostic in the editor. With the new API, an agent
client can open the same quickfix menu to trigger the "Ask Cody to fix"
command.