vscode: Relax provider search matching
This PR makes it so that all of

- github
- " GitHub"
- issues
- githubissues
- "GitHub Issues"

match `GitHub Issues`.

Previously, any excessive spaces in the query would cause all providers to not match, and you always had to put in the space. Without requiring all the spaces (but still okay to put them in) it should feel a bit more like a fuzzy finder.

Test plan:

Tim!