Cody Bench: add infrastructure for LLM Judge
Previously, we didn't have a good way to automatically measure quality
of a generated LLM response. As we are working on improving the quality
of the Fix command, this limitation was painful because we could only
use rule-based logic like "the diagnostic was fixed" that suffered from
both false positives and false negatives:

* False positives: sometimes, the LLM came up with a low-quality fix
  that commented out a bunch of code, which technically fixed the error
  but was a low-quality result.
* False negatives: sometimes, the generated fix was perfect, but it the
  generated code was missing an import (minor issue).

Using an LLM judge, we hope to address this problem. The goal is not to
rely 100% on LLM judge, but rather just use it as a tool in a larget
toolkit of eval algorithms.