vscode: Fix item count
When we switched the wording from files to items, this number made even less sense than before.

I think this was always a bit confusing as there are often more list items than there are files (because we include multiple chunks) but with the term items this just fell apart completely.

So now we no longer count the unique number of file URIs, but instead just the number of rendered items.
For telemetry, I kept it the same for now, as I am scared of skewing any metrics.

Test plan:

Verified locally that the number is now matching the number of list items.