Rewrite query for remote context (#4290)
For local keyword context (symf), we first rewrite the query to pull out
relevant keywords. Remote context does not currently rewrite. When digging into
local vs. remote context quality, we found the rewrite to be critical for (1)
queries in foreign languages, and (2) longer queries with multiple sentences or
lines.

This PR updates remote search to write the query when the question appears to
be in a foreign language or has multiple sentences. We added this restriction
because the eval results showed that rewriting made remote search much worse
for short English queries (https://github.com/sourcegraph/context-eval/pull/11)

Closes SPLF-83