commands: getContextFileFromUri treats empty ranges as full file
"Add file to cody chat" command from a right click on a file would fail
since we would have an empty range set. This treats any empty range as
an indication the user wanted to add the whole file, which should plug
the hole for other potential call sites on getContextFileFromUri.

Additionally if we do not have a range defined, we now do not set range
on the returned ContextItem. This is so we render the file in chat as
just the filename, rather than the filename with the range 1:lineCount.

Test Plan: Manually added and removed a file from chat. Then added the
file via right click on the text area, right click on the explorer.
Additionally tested that selecting text and right clicking correctly
selected just the range.