update custom command test
I don't really understand how the code I change only affects these lines
in the e2e test. But when looking at the behaviour this is better and
really anytime we get something from the active editor state it should
be the file without line ranges if there is no active selection.