isUniqueContextItem checks unset range against full range
If we are considering adding a file which is the whole file, we can
compare it to an earlier selection by comparing the content field.

Additionally we fix a bug in isUniqueContextItem where it only ever
considered the first file with the same path for deciding uniqueness.

We updated the unit test to have a correctly populated content field. In
the next commit I will add a test to make it clearer.