chat: do not auto add selection context if source is chat
We now explicitly mention the current file/selection. This bit of logic
was making it so that if we remove the mention for current file it would
still get added in if code was selected.

Test Plan: Did a chat with code selected. In one case with the @-mention
and one time without.