remove agent test chat/submitMessage understands the selected text
We now rely on explicitly mentioning the context, which means this test
doesn't really make sense. Maybe there is a way for it to be useful?

I also had to update recordings, and I'm unsure if all is good in the
world since so many recordings changed without the auto inclusion of
selected text.