indicate context used for followup messages
Previously, it was not clear which context was used in followup messages. The way it works is that context used previously in the conversation is automatically used for followups, but only *newly introduced* context is displayed. This was confusing, because it could seem like *only* that newly introduced context was used.

Now, the UI says `X new item` and expands to mention `Prior messages in this thread` for followup message context rows if new context was introduced. If no new context was introduced, there is no context row and the prior conversation context is still used (we assume this is implicit to the user).

Fix https://linear.app/sourcegraph/issue/CODY-2092/show-full-actual-context-used-for-followups