show version-updated toast above (relative not floating)
Currently the version-updated toast covers up the chat input. This is annoying and confusing. Now, it shows up above in flow. This is more prominent but less annoying IMO. And it's OK for it to be this prominent since it only shows up once every 2 weeks.

Fix https://linear.app/sourcegraph/issue/CODY-2116/move-the-cody-updated-toast-to-the-bottom-of-the-window