vscode: specify untrusted configuration "openctx.providers"
This is specified in the OpenCtx extensions, but when we ported the
support over to cody we did not specify this. This is important to
specify.

Test Plan: Made a new workspace and opened it and did not trust it to
ensure I was in restricted mode. I then successfully used cody with my
user configured providers. I added a workspace provider setting and
vscode warned me I needed to trust the workspace + the provider did not
show up in chat.