openctx: refactor exposeOpenCtxClient to async function
This allows initialization to happen concurrently with the other
promises in main when the configuration changes. The main motivation for
this change is it makes another change I want to make cleaner. View this
diff with whitespace changes turned off to see how minimal it actually
is.

Test Plan: Started up vscode and still saw OpenCtx providers in chat.