openctx: warn if sourcegraph.openctx ext is enabled
This commit will show a warning if the OpenCtx extension is installed
and tell the user to disable it. It will additionally open up the
extensions view filtered down to just the OpenCtx extension.

This isn't an ideal user experience, but after exploring the vscode APIs
and codebase I couldn't find a more surefire way.

The reason this is added is the extensions currently conflict, leading
to openctx providers in cody failing to initialize. I looked into ways
to try and gracefully detect and avoid conflict, but did not manage to
find a way.

Note: this checks at startup and on config change. I could add something
which listens for the OpenCtx extension, but that is more complicated
and the only time mention items break in cody is if it isn't the first
to call createController.

Note: exposeOpenCtxClient running on config change is a bug, since we
should only register the controller once and warn once. I will follow-up
with another commit to fix that.

Test Plan: enabled the openctx extension, then restarted the ext host
and the warning appeared.