remove separate natural-language search
The natural language search quickpick was removed. To perform a natural-language search, run a Cody chat and view the items in the context row.

This was only enabled for non-enterprise users, and it had not been worked on in a while as a separate quickpick. But the underlying search, symf, is used extensively in chat, which is now the recommended way to use this search feature.