allow creating a new chat even if a new unsent one exists
With the new chat UI that starts from the top, it feels more natural to expect to have multiple new unsent chats. Before this change, if you had a new chat that had not yet had any messages sent, clicking "New chat" would not do anything. Now it creates a new chat.

Fix https://linear.app/sourcegraph/issue/CODY-2074/allow-creating-a-new-chat-window-when-an-old-unsent-chat-window-is