openctx: update to directly use controller
Previously we passed the controller's client. However, the controller
also exposes the client API, but additionally has vscode specific error
reporting logic.

Additionally we updated the OpenCtxClient type we declare as Picking
only the fields we use.

We need to update client call sites to pass in providerUri as part of
the new option interface.