Agent: add automatically generated bindings for Java
Previously, we only generated bindings for Kotlin. The Eclipse plugin is
written in Java so this PR adds support to generate Java bindings as
well. Java and Kotlin are very similar so required diff is minimal.

I would normally not commit the Java files into the repo but I'm failing
to run the codegen tool on Windows so I'm committing it here to make it
easy to copy-paste changes.