Avoid addEnhancedContext in chat quality tests (#4548)
Unfortunately, the results of `symf` are not deterministic and can change when
data is reindexed. So when using `addEnhancedContext: true` LLM snapshots
tests, results may change from run-to-run, results in "recording not found"
errors.

This PR removes `addEnhancedContext`, and instead uses a static set of context
snippets representing the results of a context search.