vscode: remove warning for vite CJS deprecation in builds
This warning was bothering me and was pretty easy to resolve. I believe
our project needs to remain CJS for vscode extensions, so instead we can
just opt in to modules via the mts filename on vite's config.

  The CJS build of Vite's Node API is deprecated. See
  https://vitejs.dev/guide/troubleshooting.html#vite-cjs-node-api-deprecated
  for more details.

Test Plan: ran "pnpm build" in vscode and did not get the above warning.