vscode: show "Add File to Cody Chat" for remote files
In our enablements and when in a few cases we had resourceScheme ==
file. However, Cody supports remote files if the editor can get the
text. So instead we update those checks to be based on if the editor is
in focus.

Additionally for the explorer it will only show the command if the
enablement is also true, so we also check if filesExplorerFocus is true.

Finally we fix a bug where we showed the action on folders in the
explorers, which does not work.

Test Plan: Tested the action showed up when right clicking on the editor
or explorer if cody was open. Otherwise checked it was not shown.

Additionally checked the cody.menu.commands still showed up, but did not
show up on other panes (eg the extension view).