Windows Certificates: bundle roots.exe
Related to CODY-2176, could fix it.

Previously, we used the `win-ca` "fallback" method to install local
certificates. In the latest VS Code Insiders, this no longer works
and causes Cody to fail all network requests. This PR is an attempt to
fix this problem by removing the "fallback" and bundling the
`roots.exe` with the extension.