Windows: make win-ca also work with the agent
In the PR https://github.com/sourcegraph/cody/pull/4598, we fixed a bug
related to how we installed self-signed certificates on Windows. This PR
only fixed the issue for VS Code. This PR additionally fixes the problem
for the agent. For the agent, we use a slightly different
`extensionUri`. This PR adds logic to manually copy the `roots.exe` file
over to the directory we use as `extensionUri` with the agent.