editor: support all uri schemes for getTextEditorContentForFile
We already have a uri passed in and I am unsure why we converted that
into a uri which was just the filepath. I am assuming it was due to code
evolution.

In particular this makes it so that cody can understand context from
unsaved files.

Test Plan: create a new unsaved file with the word "hello world" in it.
Start a cody chat and ask cody about the file. Expect cody mentions
hello world in the response. Before this change we got an error.