CLI: move implementation to the agent project (#4653)
Fixes CODY-2536

Currently, the experimental cli lives in a separate `cli/` project. This means we can't access helper methods from the agent like starting an embedded client for faster startup.
This PR moves the implementation into the `agent/` project and removes the the old `cli/` directory while keeping `cli/README.md` unchanged so that old links to this page continue working. We can move the readme to a more permanent home later.