cody-bench: exit the process after finishing (#4656)
When running the tool locally, I noticed that it was hanging after completing the eval run. Probably, what's happening is that there's some background promise that hasn't completed running. I didn't experience this when I originally created the tool so it looks like a regression. This PR "fixes" the problem by calling process.exit(0), which is good enough to unblock usage of the tool