vscode: vendor in @openctx/linear-issues-provider
We plan on experimenting a bunch with linear issues and context. This
commit makes it easy to incorporate it outside of mentions. Additionally
if the experimental noodle flag is set, Linear Issues is shown in the
mentions dropdown.

Test Plan: with an empty openctx configuration and noodle set to true,
use linear issues.