Updated key shortcuts to latest bindings from Design (#1782)
Updated the JetBrains keyboard shortcuts according to the instructions
[here](https://linear.app/sourcegraph/document/keyboard-shortcuts-for-cody-on-jetbrains-2f8747a22530).

## Test plan

I wrote a unit test for the bindings. I would have liked to trigger the
keystrokes programmatically in the test (i.e., an integration test), and
then listened for the resulting actions to fire. However, without the
UI-robot tests, I was not able to find a way to simulate a keypress to
trigger an action.

So instead, I wrote a simpler and slightly weaker test, which looks up
all the actions of interest, gets their registered key bindings, and
compares them against the expected test values from Design. This
unfortunately doesn't exercise all the code paths -- specifically it
isn't really testing the routing for retry and for the
undo/cancel/dismiss cluster, and it's not guaranteeing that the actions
will trigger even if the bindings are registered. But it will definitely
catch whether we have the wrong bindings for the top-level actions.

I have not manually tested this on Windows yet, because I still need to
set up Parallels, and I'm headed out for PTO. But the unit test should
work on Windows as well.