clean up
This change adds a new `isMultiModalModel` method to the `ModelsService` class to detect if a given model ID represents a multimodal model. It also updates the `SimpleChatModel.addImages` method to only add images if the current model is a supported multimodal model.

The changes in the `Toolbar` component ensure that the image upload button is only shown for multimodal models.