[DRAFT] Edit: one-click edit command
This change adds a new "One Click Edit" command to the Cody extension for Visual Studio Code. The command is accessible through the Cody command menu and can be triggered with the keyboard shortcut "alt+j". This feature is marked as experimental.

The new command is registered in the `main.ts` file, and the corresponding menu item is added to the `CodyCommandMenuItems` array in the `index.ts` file. The command is enabled when the Cody extension is activated.