Windows Self-signed Certificates: use `inject: '+'`
Previously, some users were unable to run Cody on Windows with HTTPS
proxies and self-signed certificates. This PR is an attempt to fix that
problem by enabling the `inject: '+'` feature in the `win-ca` npm
package, which has the following documentation:

```
new experimental mode: `tls.createSecureContext()` is patched and
certificates are used along with built-in ones. This mode should affect
all secure connections, not just https module.
```

I manually verified that I'm able to use the plugin on the latest
Insiders with a local build of the extension. However, this might not
mean much because I'm unable to reproduce the original issue even
when I don't reference the bundled `win-ca-roots.exe` file. Either way,
this change is probably worth a shot and we can revert it if we get
user reports about network issues.