Manual Stable Release (#1862)
Our current setup produces a lot of stable releases that often are not
unhidden eventually. That makes the QA and awaiting JB approval process
parallel. Based on the recent experience only 1/8 releases goes public.
That is a waste of JB approval team's time and CI time.

Let's change the setup. The nightly version handling does not change but
the default release script does not longer publish the stable version.
Instead the separate workflow can be triggered on the specific tag that
publishes the stable version.

## Test plan
1. push a tag for a nightly release
2. trigger the stable release manually

to be tested once merged