test/fixtures: mock /.api/client-config during tests, allow SiteProductVersion
Previously, `ConfigFeaturesSingleton` would ignore all GraphQL errors
and always return a default configuration. Additionally, we did not rely
on `SiteProductVersion` in any part of the startup path. With the new
`ClientConfigSingleton` we rely on that request in the initialization
path which means the request is unauthenticated and needs to be allowed.

We also now mock `/.api/client-config` properly, as it is now required
to succeed in order to use Cody.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>